mandelbrot Fortran code: fixes warning with recent versions of gcc

+++ src/mandel.f90
@@ -12,7 +12,7 @@ subroutine mandelbrot( orig, dx, dy, R, NX, NY, NMAX )
   integer,intent(in) :: NX, NY, NMAX
   integer*2,intent(inout) :: R(NX,NY)
   complex(kind=kind(0d0)) :: point, z, zn
-  integer :: i,j,k
+  integer*2 :: i,j,k
 
   point = orig
   do i=1,NX
